requirements.txt
numpy
pickle-mixin
streamlit
streamlit-option-menu
scikit-learn
protobuf
plotly.express
